% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{as.sp}
\alias{as.sp}
\title{Converts a dataframe to a spatial points dataframe.}
\usage{
as.sp(xy)
}
\arguments{
\item{xy}{data.frame or matrix with columns named 'lat' and 'lon' or 'latitude' and 'longitude'}
}
\value{
same data.frame as a SpatialPointsDataFrame
}
\description{
Converts a dataframe to a spatial points dataframe.
}
