% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{brkfun}
\alias{brkfun}
\title{Utility function for creating breakpoints and colors for plotting}
\usage{
brkfun(v, crop = 1, sym = F, pad = T)
}
\arguments{
\item{v}{a numeric vector of values}

\item{crop}{sets the breaks and colors to cover this quantile of values.  useful if one or two outliers dominate the colorscale}

\item{sym}{logical.  if true, a blue/white/red colorscheme is used, with breaks centered on zero}
}
\value{
a list with breaks, mids, colors, and matrix used for plotting the colorscale
}
\description{
Makes a set of plots including:
map of fits
scatterplot of fits vs response
map of errors
qqnorm plot
variogram
}
