% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{dpar}
\alias{dpar}
\title{Resets all graphical parameters to default values and allows changes}
\usage{
dpar(...)
}
\arguments{
\item{...}{new parameters to pass to par()}
}
\value{
an invisible output from par() call
}
\description{
Resets all graphical parameters to default values and allows changes
}
