% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{fdates}
\alias{fdates}
\title{Extracts date strings in YYYYMMDD format from filenames. These must be separated using underscores.}
\usage{
fdates(filenames, dateformat = F, n = NULL)
}
\arguments{
\item{filenames}{a character vector, presumably filenames.}

\item{dateformat}{whether to attempt to convert these to R date format}

\item{n}{manually specifies which underscore-separated element to extract}
}
\value{
a text string or vector of dates, depending on 'dateformat'
}
\description{
Extracts date strings in YYYYMMDD format from filenames. These must be separated using underscores.
}
