% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{gdalProjRaster}
\alias{gdalProjRaster}
\title{Reprojects a DEM using an external call to the GDAL function 'gdalwarp'}
\usage{
gdalProjRaster(
  img,
  target,
  method = "bilinear",
  filename = NULL,
  wm = 1000,
  nt = 1,
  ot = "Float32",
  NAflag = -9999,
  overwrite = F
)
}
\arguments{
\item{img}{a raster object to be reprojected, or the filename for one.}

\item{target}{a raster object with the desired projection and extent.}

\item{method}{interpolation method}

\item{wm}{memory to allocate for the gdal}

\item{nt}{number of threads to use in gdal}

\item{ot}{datatype of output following gdal conventions}

\item{NAflag}{nodata value given to output}

\item{overwrite}{whether or not to overwrite an existing outname}

\item{outname}{a filename for the output.  Can be left blank.}
}
\value{
invisibly returns the reprojected rasters, and/or writes it out if 'outname' is given.
}
\description{
Reprojects a DEM using an external call to the GDAL function 'gdalwarp'
}
