% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{get_proj_string}
\alias{get_proj_string}
\title{Writes out a system call to gdal_translate or gdalwarp ~~~~~~~}
\usage{
get_proj_string(img, tgt, band = 1)
}
\arguments{
\item{img}{a rast raster}

\item{locs}{a data.frame with columns named 'lat' and 'lon'}

\item{...}{additional arguemnts to terra::extract e.g. method='bilinear'}
}
\value{
numeric vector of values at points
}
\description{
Writes out a system call to gdal_translate or gdalwarp ~~~~~~~
}
