% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{gwarp}
\alias{gwarp}
\title{reprojects a raster using gdal commands ~~~~~~~}
\usage{
gwarp(img, tgt, fn.out, execute = T)
}
\arguments{
\item{img}{a spatRaster object or the name of a raster file}

\item{tgt}{a spatRaster object, the name of a raster file, or a numeric vector defining an extent.}

\item{fn.out}{a filename for the output raster}

\item{execute}{logical, whether to execute the command or return it as a string}
}
\value{
either a spatRaster object invisiblly, or a text string with a gdal system command
}
\description{
reprojects a raster using gdal commands ~~~~~~~
}
