% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{gwr.coef.fit-class}
\alias{gwr.coef.fit-class}
\title{An S4 class to for a raster of GWR coefficients}
\description{
An S4 class to for a raster of GWR coefficients
}
\section{Slots}{

\describe{
\item{\code{tmplt.name}}{filename of a raster grid over which to fit GWR coefficient values}

\item{\code{coef.name}}{filename of the coefficient raster}

\item{\code{band.name}}{character vector of names for each of the bands of the coefficient raster}

\item{\code{hillshade.name}}{filename of a hillshade raster with same extent as the coefficient raster}
}}

