% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwr_loocv.R
\name{gwr_loocv}
\alias{gwr_loocv}
\title{Fits a GWR regression to tabular data contained in a data.frame}
\usage{
gwr_loocv(
  form,
  data,
  name = "",
  bw = NaN,
  adaptive = F,
  idx = NULL,
  n.samp = 3000,
  verbose = T,
  buffer = 3,
  fast = T
)
}
\arguments{
\item{form}{a standard R formula in character format e.g. z~x+y}

\item{name}{a character vector used in any filenames created.}

\item{bw}{fixed bandwidth in km.  Can be left blank if 'abw' is set.  Error will be triggered if both 'bw' and 'abw' are set.}

\item{idx}{optional vector of row numbers at which to perform LOOCV.}

\item{n.samp}{optional max number of locations at which to perform LOOCV.}

\item{rdata}{a data.frame containing the variables in 'form', along with columns named 'lat' and 'lon'.}

\item{out.dir}{a filepath for outputs}

\item{abw}{adaptive bandwidth size.  Sets bandwidth as radius required to capture n neighbors.  Can be left blank if 'bw' is specified.}
}
\value{
a gwr object with the original inputs, original data, fit, statistics.
}
\description{
This function fits a gwr model to a data.frame which includes geographic information
}
