% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{hillshade}
\alias{hillshade}
\title{Makes a hillshade raster from a dem raster}
\usage{
hillshade(dem)
}
\arguments{
\item{dem}{a digital elevation model in raster format.}
}
\value{
a hillshade in raster format.
}
\description{
Makes a hillshade raster from a dem raster
}
