% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{listfiles}
\alias{listfiles}
\title{wrapper for list.files() with additional features ~~~~~~~
'pattern' can be a vector
'any'
'strict'
'none'}
\usage{
listfiles(..., include.dirs = F, full = T)
}
\arguments{
\item{...}{additional arguemnts to terra::extract e.g. method='bilinear'}
}
\value{
numeric vector of values at points
}
\description{
wrapper for list.files() with additional features ~~~~~~~
'pattern' can be a vector
'any'
'strict'
'none'
}
