% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ps.r
\name{make_ps}
\alias{make_ps}
\title{Makes a pansharpener raster, which is the ratio of a finescale fit divided by a smoothed version of the same fit.
This is accomplished by aggregating a}
\usage{
make_ps(
  gwr_fine,
  smooth.fact = 8,
  verbose = T,
  overwrite = F,
  overwrite.smooth = F,
  resamp.method = "bilinear",
  do.stats = F
)
}
\arguments{
\item{verbose}{logical; whether or not to report progress}

\item{overwrite}{logical; whether or not to overwrite existing raster fit}

\item{img.coarse}{a filename for a coarse-res prediction}

\item{img.fine}{filename for a fine-res prediction}
}
\value{
a gwr object with the original inputs, original data, fit, statistics.
}
\description{
This function is for pan-sharpening a more accurate coarse grid using finer, but possibly biased, estimate of the same thing.
Based on ratio compnenent substitucion (RCS) pan-sharpening following
Mhangara, P., Mapurisa, W. and Mudau, N., 2020. Comparison of image fusion techniques using satellite pour l’Observation
  de la Terre (SPOT) 6 satellite imagery. Applied Sciences, 10(5), p.1881.
see 'https://www.r-bloggers.com/2015/02/pan-sharpening-using-r/'
Steps:
1. resample the coarse image to the fine resolution
2. resample the fine image to the coarse resolution
3. resample the coarse version of the fine back to fine res.
4. compute ratio of the original fine imag to it's smoothed version (3)
5. multiply (1) by (4)
}
