% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{make_subset}
\alias{make_subset}
\title{crops a set of spatRaster images ~~~~~~~}
\usage{
make_subset(
  my.ext,
  fn,
  dir = "/mnt/DataDrive1/data/wg/raster/30m_subsets/",
  ext.name = NULL
)
}
\arguments{
\item{dir}{an existing directory.  a subdirectory based on the center of the extent will be created}

\item{img}{a spatRaster object or the name of a raster file}

\item{ext}{a spatRaster extent, or a numeric vector describing an extent.}
}
\value{
a vector of filenames for the subset rasters.
}
\description{
crops a set of spatRaster images ~~~~~~~
}
