% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{objs}
\alias{objs}
\title{lists objects in the current environment, and their size ~~~~~~~}
\usage{
objs(n = 10, verbose = T, envir = .GlobalEnv)
}
\arguments{
\item{...}{additional arguemnts to terra::extract e.g. method='bilinear'}
}
\value{
numeric vector of values at points
}
\description{
lists objects in the current environment, and their size ~~~~~~~
}
