% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pan_sharpen.R
\name{pan_sharpen}
\alias{pan_sharpen}
\title{Makes GWR coefficient estimates on a grid}
\usage{
pan_sharpen(
  gwr_coarse,
  gwr_fine,
  smooth.fact = 8,
  verbose = T,
  overwrite = F,
  overwrite.smooth = F,
  suffix = "pansharpen",
  gdal = T
)
}
\arguments{
\item{verbose}{logical; whether or not to report progress}

\item{overwrite}{logical; whether or not to overwrite existing raster fit}

\item{img.coarse}{a filename for a coarse-res prediction}

\item{img.fine}{filename for a fine-res prediction}
}
\value{
a gwr object with the original inputs, original data, fit, statistics.
}
\description{
This function is for pan-sharpening a more accurate coarse grid using finer, but possibly biased, estimate of the same thing.
Steps:
1. resample the coarse image to the fine resolution
2. resample the fine image to the coarse resolution
3. resample the coarse version of the fine back to fine res.
4. compute ratio of the original fine imag to it's smoothed version (3)
5. multiply (1) by (4)
}
