% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot,gwr,ANY-method}
\alias{plot,gwr,ANY-method}
\title{Diagnostic plots of a LOOCV GWR fit.}
\usage{
\S4method{plot}{gwr,ANY}(x, title = NULL, max.pts = 7500)
}
\arguments{
\item{x}{a gwr object created using gwr_loocv()}
}
\value{
nothing
}
\description{
Makes a set of plots including:
map of fits
scatterplot of fits vs response
map of errors
qqnorm plot
variogram
}
