% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot,gwr.coef.fit,ANY-method}
\alias{plot,gwr.coef.fit,ANY-method}
\title{Coefficient plots for a GWR fit.}
\usage{
\S4method{plot}{gwr.coef.fit,ANY}(x, hill = NULL, crop = 0.99, title = NULL)
}
\arguments{
\item{x}{a gwr.coef.fit object created using gwr_coef_map()}
}
\value{
nothing
}
\description{
Makes a set of plots for each cofficient of a gwr model
}
