% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot,gwr.raster.fit,ANY-method}
\alias{plot,gwr.raster.fit,ANY-method}
\title{Plots for a GWR raster fit.}
\usage{
\S4method{plot}{gwr.raster.fit,ANY}(x, title = NULL, mar.fract = 0.18, show.errs = F, zlab = NULL, ...)
}
\arguments{
\item{x}{a gwr.coef.fit object created using gwr_coef_map()}
}
\value{
nothing
}
\description{
Makes a set of plots for each cofficient of a gwr model
}
