% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot,RasterLayer,ANY-method}
\alias{plot,RasterLayer,ANY-method}
\title{plot a raster layer}
\usage{
\S4method{plot}{RasterLayer,ANY}(x, hill = NULL, sym = F, crop = 1, zlab = "", mar.fract = 0.15, ...)
}
\arguments{
\item{x}{a RasterLayer object}
}
\value{
nothing
}
\description{
Makes a set of plots including:
map of fits
scatterplot of fits vs response
map of errors
qqnorm plot
variogram
}
