% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{pplot}
\alias{pplot}
\title{plot a SpatRaster layer}
\usage{
pplot(
  x,
  hill = NULL,
  sym = F,
  crop = 1,
  zlab = "",
  mar.fract = 0.18,
  brks = NULL,
  ...
)
}
\arguments{
\item{x}{a RasterLayer object}
}
\value{
nothing
}
\description{
Makes a set of plots including:
map of fits
scatterplot of fits vs response
map of errors
qqnorm plot
variogram
}
