% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{rsq}
\alias{rsq}
\title{Calculates R-squared}
\usage{
rsq(y, yh)
}
\arguments{
\item{y}{response varaible}

\item{yh}{prediction}
}
\value{
numeric R-squared value
}
\description{
This function tests to see if a dummy function can be executed over a doParallel cluster using
'foreach() %dopar% ...' syntax.  Returns T or F based on whether the foreach executed without
error.  If verbose=T, the function prints the number of cores detected.
}
