% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{search.code2}
\alias{search.code2}
\title{Searches code directory for certain text strings and/or filenames.}
\usage{
search.code2(
  text = NULL,
  name = NULL,
  dir = "~/code/",
  n = 30,
  sort = c("newest", "random", "oldest", "line", "latest"),
  include = "all",
  mtl = 220
)
}
\arguments{
\item{text}{text string to search for}

\item{name}{text to search filenames for}

\item{dir}{directory to look in}

\item{n}{number of instances to show}

\item{sort}{how to sort instances.}
}
\value{
if a 'text' argument is given, prints a list of instances sorted as specified.  If a 'name' arg is given, prints filenames sorted based on modification times
}
\description{
Searches code directory for certain text strings and/or filenames.
}
