% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_cluster.R
\name{test_cluster}
\alias{test_cluster}
\title{Test for a vailid doParallel cluster}
\usage{
test_cluster(verbose = T)
}
\arguments{
\item{verbose}{whether or not to print cluster size}
}
\value{
logical; whether or not the foreach executes successfully over a cluster
}
\description{
This function tests to see if a dummy function can be executed over a doParallel cluster using
'foreach() %dopar% ...' syntax.  Returns T or F based on whether the foreach executed without
error.  If verbose=T, the function prints the number of cores detected.
}
