% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{textract}
\alias{textract}
\title{Wrapper for terra::extract which automatically reprojects points to match the raster ~~~~~~~}
\usage{
textract(img, locs, ...)
}
\arguments{
\item{img}{a rast raster}

\item{locs}{a data.frame with columns named 'lat' and 'lon'}

\item{...}{additional arguemnts to terra::extract e.g. method='bilinear'}
}
\value{
numeric vector of values at points
}
\description{
Wrapper for terra::extract which automatically reprojects points to match the raster ~~~~~~~
}
