% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{xExtent}
\alias{xExtent}
\title{Projects a geographic extent}
\usage{
xExtent(x, to, from = NULL)
}
\arguments{
\item{x}{an extent or a raster object from which to obtain an extent}

\item{to}{a new projection, or a raster object from which to obtain a projection}

\item{from}{the old projection.  defaults to the projection of x, if x is a raster object, or geographic}
}
\value{
a list with the corner coordinates in the new projection, an extent object in the new projection that contains the original extent, and a numeric vector of the same.
}
\description{
Projects a geographic extent
}
